% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{AnnotationSite}
\alias{AnnotationSite}
\title{Annotate pA site based gtf file.}
\usage{
AnnotationSite(
  pAsite,
  gtfFile,
  genome_ver,
  annotLevels = c("3UTRs", "Exon", "Intron", "CDS", "LastExon1Kb", "3UTRs_1kb",
    "3UTRs_2kb", "5UTR", "INTERGENIC"),
  cores = 1
)
}
\arguments{
\item{pAsite}{A list of pA sites, `1:810052:+`}

\item{gtfFile}{The standard gtf file.}

\item{genome_ver}{The version of genome annotation. Hg38, Hg19, Mm10, Mm9.}

\item{annotLevels}{The priority of annotaiton.}

\item{cores}{The num of cpu for DE test.}
}
\description{
Annotate pA site based gtf file.
}
