% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialTest.R
\name{FindDE}
\alias{FindDE}
\title{Find differentail expression APA events}
\usage{
FindDE(
  obj,
  idents.1,
  idents.2 = NULL,
  annot,
  paLoc = "3UTRs",
  slot = "counts",
  assay = "RNA",
  seedUse = 1,
  cores = 1,
  num.splits = 6
)
}
\arguments{
\item{obj}{A Seurat object which contained apa matrix information.}

\item{idents.1}{A sub-cluster from Idents(dat)}

\item{idents.2}{A sub-cluster from Idents(dat). If idents.2 == NULL, then idents.1 versus other all cells.}

\item{annot}{The annotation of all pA sites, generate from AnnotationSite.}

\item{paLoc}{The genomic annotation of pA for performing DE test.}

\item{slot}{The slot data of Seurat, including counts, data and scale.data. defualt: counts}

\item{assay}{The name of assay in Seurat class, default: RNA}

\item{seedUse}{The random seed used.}

\item{cores}{The num of cpu for DE test.}

\item{num.split}{The num of cell chunks for random sample from all cells.
#'@example
\dontrun{test <- FindDE(
dat,
idents.1 = levels(Idents(dat))[1],
idents.2 = levels(Idents(dat))[2],
annot = pa_anno,
assay = "RNA",
cores = 20
)
}}
}
\description{
Find DE APA events by DEXseq test
}
