% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Loading datasets}
\usage{
loadData(fileList, collapsePa, cores = 1)
}
\arguments{
\item{fileList}{A named vector of apa matrix file.}

\item{collapsePa}{The output of group_pa.py}

\item{cores}{The num of cpu for processing data.}
}
\description{
loading apa matrix into a Seurat object.
}
