% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiCategory.R
\name{psi}
\alias{psi}
\title{Psi calculation.}
\usage{
psi(obj, annot, chunk = 1000, assay = "apapsi", cores = 1)
}
\arguments{
\item{obj}{A Seurat object include `apa` assay.}

\item{annot}{Annotation dataframe.}

\item{chunk}{Cell number per chunk for reducing memory usage.}

\item{cores}{The num of cpu for calculating psi.}
}
\description{
Psi calculation.
}
